+++ CHANGES.md
@@ -39,6 +39,7 @@ Changes in 13.0
 - Fix SQL roll "Block" Periods correctly in Rollover.php
 - Add Before delete Course Period hook in Courses.php & Actions.php
 - Remove deprecated `_myURLEncode()` function in PreparePHP_SELF.fnc.php
+- SQL use CONCAT() so string is escaped if needed in GetStuList.fnc.php, Search.fnc.php & Requests.php
 
 Changes in 12.9.2
 -----------------